hw_many/messenger/posts/serializers.py
from posts.models import Post
from rest_framework import serializers


class PostSerializer(serializers.ModelSerializer):
    author = serializers.CharField(read_only=True)

    class Meta:
        model = Post
        fields = ['id', 'title', 'text', 'time', 'author']